requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
httpx==0.26.0
pydantic==2.5.3
pydantic-settings==2.1.0
pyyaml==6.0.1
tiktoken==0.5.2
python-dotenv==1.0.0
